requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.13.4
colorama==0.4.6
discord.py==2.5.2
keyboard==0.13.5
matplotlib==3.10.1
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
Requests==2.32.3
rich==14.0.0
selenium==4.31.0
webdriver_manager==4.0.2
yfinance==0.2.40
